+++ public/index.html
@@ -2,19 +2,19 @@
 <html lang="en">
 
 <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
-    <meta name="theme-color" content="#000000" />
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-    <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico" />
-    <title>React App</title>
+	<meta charset="utf-8" />
+	<meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
+	<meta name="theme-color" content="#000000" />
+	<link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
+	<link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico" />
+	<title>React App</title>
 </head>
 
 <body>
-    <noscript>
-        You need to enable JavaScript to run this app.
-    </noscript>
-    <main id="root"></main>
+	<noscript>
+		You need to enable JavaScript to run this app.
+	</noscript>
+	<main id="root"></main>
 </body>
 
 </html>
