+++ tsconfig.prod.json
@@ -1,3 +1,3 @@
 {
-    "extends": "./tsconfig.json"
+	"extends": "./tsconfig.json"
 }
